Improve kill zones.

+++ src/kill_zone.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Clone)]
 pub struct KillZone {
-    pub start_time: f32,
-    pub end_time: f32,
+    pub start_time: usize,
+    pub end_time: usize,
     pub position: (u32, u32),
     pub width: u32,
     pub height: u32,
@@ -40,7 +40,7 @@ pub fn distance_to_killzone(active_kill_zones: &Vec<&KillZone>, (x, y): (u32, u3
 pub fn is_point_in_killzone(
     kill_zones: &Vec<KillZone>,
     (x, y): (u32, u32),
-    generation_time: f32,
+    generation_time: usize,
 ) -> bool {
     kill_zones.iter().any(|kz| {
         generation_time >= kz.start_time
